- almost works, just needs to withdrawal from transactions - pretty close, the ledger seems to work - just missing history from poloniex? - seems to work well, but the lending has messed things up - fixes and logging (WORKS) - Add "event" support
Whereas an "event" is the pairing of 1 or more source transactions with
a "destination" transaction. Essentially, it would constitute the
complete act of buying and selling (and realizing gains or losses).

IANAA (I am not an accountant) so this may or may not be a correct
implementation, but looks pretty snazzy.